% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_rma.R
\name{ow_rma}
\alias{ow_rma}
\title{Computation of one-way repeated measures ANOVA (rma)}
\usage{
ow_rma(ow_rma_data)
}
\arguments{
\item{ow_rma_data}{datataframe with subject id column and k factor columns.}
}
\value{
ANOVA Table
\item{ow_rma_data}{The first object input.}
}
\description{
Estimate an .repeated measurement on fiven data
}
\note{
This is the main function of the MAGA-package to estimate the repeated measurement ANOVA.
}
\examples{

ANOVA_table = ow_rma(some_rma_data)

}
\author{
Sums of Suares
}

