% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_rma_ci.R
\name{ow_rma_ci}
\alias{ow_rma_ci}
\title{Computation of adjusted and unadjusted confidence intervalls for one-way ANOVA with repeated measurement}
\usage{
ow_rma_ci(ow_rma_data)
}
\arguments{
\item{ow_rma_data}{datataframe with subject id column and k factor columns.}
}
\value{
list with adj. and unadj. confidence intervals
\item{ow_rma_data}{The first object input.}
}
\description{
Estimate adjusted and unadjusted confidence intervals for repeated measurement ANOVA.
}
\note{
This functions allows to calculate the confidence intervals of a repeated measurement ANOVA.
}
\examples{

CI_list = ow_rma_ci(some_rma_data)

# adjusted CIs
CI_list$adjusted_CI

# unadjusted CIs
CI_list$unadjusted_CI


}
\author{
Sums of Suares
}

