% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_rma_eta.R
\name{ow_rma_eta}
\alias{ow_rma_eta}
\title{Effect size measures for a one-way repeated measures ANOVA}
\usage{
ow_rma_eta(ow_rma_data, append = FALSE)
}
\arguments{
\item{ow_rma_data}{An object of type data.frame. Each row should represent one subject and each column one variable.}

\item{append}{Logical. If FALSE only the effect sizes. If TRUE, an ANOVA table augmented with effect sizes is returned aditionally. Default is FALSE.}

\item{id}{An integer specifying the column position of the subject ID. Default is 1. Set to "none" if the data does not contain an ID variable.}
}
\value{
Returns an object of type list.
\item{effect_size_table}{An object of type data.frame containing the effect sizes eta square and partial eta square}
\item{anova_table}{An object of type data.frame. Contains the anova table augmented with the effect size measures}
}
\description{
Compute eta-square and partial eta_square for repeated measures ANOVA. Partial eta square is the effect size for the total variance reduced by the inter-subject variance (as done in a repeated measures ANOVA model)
}
\note{
LITERATURE
}
\examples{
CI_list = ow_rma_eta(some_rma_data)



}
\author{
Joachim Munch, Frederik Schreck, Quang Nguyen Duc, Constantin Meyer-Grant, Nikolas Hoeft
}

