% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_rma_sse_reduct.R
\name{ow_rma_sse_reduct}
\alias{ow_rma_sse_reduct}
\title{Visualize reduction of SSE due to model selection}
\usage{
ow_rma_sse_reduct(ow_rma_data)
}
\arguments{
\item{ow_rma_data}{An object of type data.frame. Each row should represent one subject and each column one variable.}

\item{id}{An integer specifying the column position of the subject ID. Default is 1. Set to "none" if the data does not contain an ID variable.}

\item{plot_type}{A character specifying the type of plot that is returned to visualize the error variance reduction. Possible values are “pie” and “bar”. Default is “pie”.}

\item{return_anova_table}{Logical. If TRUE, a regular oneway ANOVA table without repeated measurements is returned additionally. Default is FALSE.}
}
\value{
Returns an object of type list.
\item{plot}{A ggplot object. The type of plot is determined by the argument plot_type}
\item{anova_table}{An object of type data.frame containing a regular oneway ANOVA table without repeated measurements}
}
\description{
Compare the sums of squared errors in regular one-way ANOVA and one-way RM ANOVA.
}
\note{
Note that the one-way ANOVA without repeated measures is for illustration purposes only since the data structure is correlated across the factor levels because of the dependent measurements.The ANOVA without repeated measures treats the data as if they were independend i.e. as if there were different subjects in each group, which is in fact not the case.
}
\examples{


}
\author{
Joachim Munch, Frederik Schreck, Quang Nguyen Duc, Constantin Meyer-Grant, Nikolas Hoeft
}

