% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_rma_ci.R
\name{rma_ci}
\alias{rma_ci}
\title{Compute of adjusted and unadjusted confidence intervals for one-way repeated measures ANOVA}
\usage{
rma_ci(rma_data, C_level = 0.95, id = 1, print_plot = TRUE)
}
\arguments{
\item{rma_data}{An object of type data.frame that contains the data for the repeated measures ANOVA. Each row should represent one entity and each column one variable.}

\item{C_level}{= A numeric value specifying the level of the confidence intervals. Default is 0.95.}

\item{id}{An integer specifying the column position of the subject ID. Default is 1. Set to "none" if the data does not contain an ID variable.}

\item{print_plot}{Logical. If TRUE, function returns a plot that displays unadjusted vs. adjusted confidence intervals. Default is TRUE.}
}
\value{
Returns an object of type list.
\item{ci_plot}{A ggplot object displaying the comparison of adjusted and unadjusted confidence intervals}
\item{confidence_intervals}{An object of type data.frame. Contains adjusted and unadjusted confidence intervals of specified level.}
}
\description{
Estimate adjusted and unadjusted confidence intervals for repeated measurement ANOVA. Adjusted confidence intervals take into account the inter-subject variance. The functions allows to visually compare both types.
}
\note{
References:

Morey, R. D. (2008). Confidence intervals from normalized data: A correction to cousineau (2005), reason 4(2): 61–64.

O’Brien, F., & Cousineau, D. (2014). Representing error bars in within-subject designs in typical software packages. The Quantitative Methods for Psychology, 10(1), 56–67.
}
\examples{

CI_list = rma_ci(some_rma_data, C_level = .95, id = 1, print_plot = TRUE)

# adjusted CIs
CI_list$adjusted_CI

# unadjusted CIs
CI_list$unadjusted_CI


}
\author{
Joachim Munch, Frederik Schreck, Quang Nguyen Duc, Constantin Meyer-Grant, Nikolas Hoeft
}

