% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_rma_spheri.R
\name{rma_spheri}
\alias{rma_spheri}
\title{Test for sphericity and p-value adjustment}
\usage{
rma_spheri(rma_data, id = 1, append = FALSE)
}
\arguments{
\item{rma_data}{An object of type data.frame. Each row should represent one subject and each column one variable.}

\item{id}{An integer specifying the column position of the subject ID. Default is 1. Set to "none" if the data does not contain an ID variable.}

\item{append}{Logical. If FALSE only the results from the sphericity test are returned along with three different p-value adjustments. If TRUE, an ANOVA table with appropriately adjusted p-values is returned aditionally. Default is FALSE.}
}
\value{
Returns an object of type list.
\item{mauchly_table}{An object of type data.frame containing the results from the Mauchly test for sphericity}
\item{correction_factors}{An object of type data.frame containing the correction factors for p-values in the presence of sphericity}
\item{corrected_rm_anova}{An object of type data.frame. RM ANOVA table with appropriately adjusted p-values}
}
\description{
Perform a Mauchly test for sphericity and compute p-value adjustments
}
\note{
LITERATURE REFERENCES
}
\examples{

rma_spheri(some_rma_data, id = 1, append = FALSE)

}
\author{
Joachim Munch, Frederik Schreck, Quang Nguyen Duc, Constantin Meyer-Grant, Nikolas Hoeft
}

