% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ow_rma_data.R
\name{sim_rma_data}
\alias{sim_rma_data}
\title{Simulate data for one-way repeated measures ANOVA}
\usage{
sim_rma_data(n, k, means = NULL, poly_order = NULL, noise_sd = 10,
  between_subject_sd = 40, NAs = 0)
}
\arguments{
\item{n}{Integer specifying the number of subjects.}

\item{k}{Integer specifying the number of factor levels i.e. repeated measurements.}

\item{means}{Numeric vector specifying the means for the k factor levels. If length of means vector exeeds k, the mean vector is truncated to match the number of factor levels. Default is NULL.}

\item{poly_order}{= Integer specifying the highest order polynomial trend that the simulated data is supposed to have. Ignored if means argument is specified. Default is NULL.}

\item{noise_sd}{Numeric vector specifying the standard deviation of the noise that is added to each factor level values. If it is desired to simulate sphericity, noise_sd must be of length k, otherwise of length 1. Default is 10.}

\item{between_subject_sd}{Numeric vector specifying the standard deviation of the values withing each of the k factor levels. Default is 40.}

\item{NAs}{Integer specifying the number of missing values that is randomly assigned to the data. Only used to ensure that the functions are robust against missing values. Default is 0.}
}
\value{
Returns an object of type data.frame
\item{rma_data}{An object of type data.frame containing the simulated data}
}
\description{
Simulate data that can be used to estimate repeated measures ANOVA models. Function parameters allow to simulate different data properties such as sphericity and polynomial trends.
}
\note{
This function allows to generate custom rma data that is suitable for all MAGA functions.
}
\examples{

sim_rma_data(20, 4, means = NULL, poly_order = NULL, noise_sd = 10, between_subject_sd = 40, NAs = 0)

}
\author{
Joachim Munch, Frederik Schreck, Quang Nguyen Duc, Constantin Meyer-Grant, Nikolas Hoeft
}

