% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_rma_spheri.R
\name{rma_spheri}
\alias{rma_spheri}
\title{Test for sphericity and p-value adjustment}
\usage{
rma_spheri(rma_data, id = 1, append = FALSE)
}
\arguments{
\item{rma_data}{An object of type data.frame. Each row should represent one subject and each column one variable.}

\item{id}{An integer specifying the column position of the subject ID. Default is 1. Set to "none" if the data does not contain an ID variable.}

\item{append}{Logical. If FALSE only the results from the sphericity test are returned along with three different p-value adjustments. If TRUE, an ANOVA table with appropriately adjusted p-values is returned aditionally. Default is FALSE.}
}
\value{
Returns an object of type list.
\item{mauchly_table}{An object of type data.frame containing the results from the Mauchly test for sphericity}
\item{correction_factors}{An object of type data.frame containing the correction factors for p-values in the presence of sphericity}
\item{corrected_rm_anova}{An object of type data.frame. RM ANOVA table with appropriately adjusted p-values}
}
\description{
Perform a Mauchly test for sphericity and compute p-value adjustments
}
\note{
References:

Geisser, S., Greenhouse, S. W. (1958). An extension of box’s results on the use of the f distribution in multivariate analysis, The Annals of Mathematical Statistics 29(3): 885–891.

Greenhouse, S. W., & Geisser, S. (1959). On methods in the analysis of profile data. Psychometrika, 24(2), 95–112.

Huynh, H., & Feldt, L. S. (1970). Conditions under Which Mean Square Ratios in Repeated Measurements Designs Have Exact F-Distributions. Journal of the American Statistical Association, 65(332), 1582–1589.

Huynh, H., & Feldt, L. S. (1976). Estimation of the Box Correction for Degrees of Freedom from Sample Data in Randomized Block and Split-Plot Designs. Journal of Educational and Behavioral Statistics, 1(1), 69–82.

Mauchly, J. W. (1940). Significance test for sphericity of a normal n-variate distribution. The Annals of Mathematical Statistics, 11(2), 204–209.
}
\examples{

rma_spheri(some_rma_data, id = 1, append = FALSE)

}
\author{
Joachim Munch, Frederik Schreck, Quang Nguyen Duc, Constantin Meyer-Grant, Nikolas Hoeft
}

